backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zt_ate_interview_agent"
version = "0.1.0"
description = "Zero-Trust Autonomous Talent Ecosystem Interview Agent"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "aiohttp>=3.10.0",
  "aiosqlite>=0.20.0",
  "PyJWT[cryptography]>=2.9.0",
  "pydantic>=2.8.0,<3.0.0",
  "pydantic-settings>=2.4.0",
  "langgraph>=0.2.27",
  "typing-extensions>=4.12.0",
  "orjson>=3.10.0",
  "google-generativeai>=0.8.0"
]

[project.optional-dependencies]
full = [
  "pydantic-ai>=0.0.20",
  "crewai>=0.95.0",
  "agno>=1.4.0"
]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "httpx>=0.27.0",
  "ruff>=0.6.0"
]

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["data*", "tests*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "ASYNC", "B"]
